TATA FORAGE
Hey Guys 
The TATA virtual Internship is offered via the FORAGE platform. It helps to get a good insight into Framing Relevant questions on behalf of a business leader. Choose the right visualization for your charts and later on convey the insights you gain from these works. It's a great opportunity if you are looking forward to learning things as a data analyst and have a keen interest in business.